requirements.txt
# fastapi>=0.110.0
# uvicorn[standard]>=0.23.0
# streamlit>=1.31.0
# langgraph>=0.2.0
# langchain>=0.2.0
# langchain-community>=0.2.0
# langchain-openai>=0.1.0
# langchain-chroma>=0.1.0
# langchain-text-splitters>=0.0.1
# chromadb>=0.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
#google-generativeai>=0.1.0
google-genai>=0.1.0